requirements.txt
Flask==0.11
Flask-Script==2.0.5
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11.10
click==6.6
itsdangerous==0.24
pymongo==3.2.2
wsgiref==0.1.2
